+++ CHANGELOG.md
@@ -15,6 +15,11 @@ It documents the changes in each of the tagged releases
 + Created `PSWSMan` which is a PowerShell module uploaded to the PowerShell Gallery that can install and manage the OMI libraries for you
 + Build `libpsrpclient` as well and add it to the release artifacts
 + Added Alpine 3 to the build matrix
++ Added support for reading `New-PSSessionOption -SkipCACheck -SkipCNCheck` from PowerShell instead of relying on the env vars
+  + Requires PowerShell v7.2.0
+  + v7.2.0 and later do not need to have `-SessionOption (New-PSSessionOption -SkipCACheck -SkipCNCheck)` set
+  + Those options can now also control cert verification behaviour per session
+  + Older versions must still set those session options and use the env vars to skip cert verification
 
 ## 1.2.1 - 2020-09-26